fotoupload met of zonder maakt ni uit

+++ app/Http/Controllers/ProjectController.php
@@ -285,7 +285,8 @@ class ProjectController extends Controller
     public function postEditProject($project_id)
     {
         $project = Project::find($project_id);
-
+        $oldimage = $project['img'];
+        $oldtricolor = $project['img_tricolor'];
         $title = Request::input('title');
         $body = Request::input('body');
         $tags = Request::input('tags');
@@ -300,7 +301,7 @@ class ProjectController extends Controller
             'tags' => $tags,
         );
         $rules = array(
-            'fileToUpload' => 'required|image',
+            'fileToUpload' => 'image',
             'title' => 'required',
             'body' => 'required',
             'tags' => 'required',
@@ -317,15 +318,19 @@ class ProjectController extends Controller
 
         //if input is a file upload to destination /uploads
         //else redirect to view with error message
-        if($image->isValid()){
-            $extension = $image->getClientOriginalExtension(); // getting image extension
-            $fileName = Auth::user()->username . '_' . rand(11111,99999).'.'.$extension; // renaming image
-            File::delete('uploads/' . $project->img);
-            $img = Image::make($image);
-            $img->fit(500,500)->crop(500, 500, 0, 0)->save($destinationPath . $fileName);
-        } else {
-            Session::flash('error', 'uploaded file is not valid');
-            return Redirect::to('/projects/' . $project_id . '/edit');
+        if($image == ""){
+            $fileName = $oldimage;
+        }else {
+            if($image->isValid()){
+                $extension = $image->getClientOriginalExtension(); // getting image extension
+                $fileName = Auth::user()->username . '_' . rand(11111,99999).'.'.$extension; // renaming image
+                File::delete('uploads/' . $project->img);
+                $img = Image::make($image);
+                $img->fit(500,500)->crop(500, 500, 0, 0)->save($destinationPath . $fileName);
+            } else {
+                Session::flash('error', 'uploaded file is not valid');
+                return Redirect::to('/projects/' . $project_id . '/edit');
+            }
         }
 
         $project->title = $title;
@@ -333,19 +338,22 @@ class ProjectController extends Controller
         $project->tags = $tags;
         $project->img = $fileName;
 
-        if($extension == 'png'){
-            $pimage = $client->loadPng('uploads/' . $fileName);
-        } else if($extension == 'jpg'){
-            $pimage = $client->loadJpeg('uploads/' . $fileName);
-        } else{
-            return 'wrong image';
+        if($image == ""){
+            $tricolor_correct = $oldtricolor;
+        }else {
+            if($extension == 'png'){
+                $pimage = $client->loadPng('uploads/' . $fileName);
+            } else if($extension == 'jpg'){
+                $pimage = $client->loadJpeg('uploads/' . $fileName);
+            } else{
+                return 'wrong image';
+            }
+
+            $image_tricolor = $pimage->extract(3);
+            $tricolor_correct = str_replace('#', '', "".$image_tricolor[0].",".$image_tricolor[1].",".$image_tricolor[2]);
         }
 
-        $image_tricolor = $pimage->extract(3);
-        $tricolor_correct = str_replace('#', '', "".$image_tricolor[0].",".$image_tricolor[1].",".$image_tricolor[2]);
         $project->img_tricolor = $tricolor_correct;
-
-
         $project->user_id = Auth::id();
         //$project->user_id = 1;
         //$table->increments('id');
